Factor out code in ElixirDelhiBot into ElixirDelhiBot.Greeter
ElixirDelhiBot knows about Greeter, Greeter knows how to greet.
